% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformMeasures.R
\name{Acc}
\alias{Acc}
\title{Accuracy}
\usage{
Acc(ct)
}
\arguments{
\item{ct}{Confusion Matrix}
}
\value{
Accuracy
}
\description{
Accuracy
}
\examples{
diag <- as.numeric(speMGX[,1])
diag[diag == 3] <- 1  # De 3 a 2 classes: No Malalt /  malalt
CM <- classify(data=speMGX[,7:ncol(speMGX)],y=diag,kernel="qJac")
Acc(CM)
}
