% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernFunct.R
\name{aitch.dist}
\alias{aitch.dist}
\title{Aitchison distance kernel}
\usage{
aitch.dist(data)
}
\arguments{
\item{data}{A matrix or data.frame containing only positive values.}
}
\value{
Aitchison distance matrix
}
\description{
Aitchison distance kernel
}
\examples{
example <- matrix(abs(rnorm(12)),nrow=4,ncol=3)
aitch.dist(data=example)
}
