% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernFunct.R
\name{aitch}
\alias{aitch}
\title{Kernel clr versió cutre}
\usage{
aitch(data)
}
\arguments{
\item{data}{A matrix or data.frame containing only positive values.}
}
\value{
Aitchison matrix
}
\description{
1 - aitchison distance / max(aitchison distance)
}
\examples{
example <- matrix(abs(rnorm(12)),nrow=4,ncol=3)
qJaccMatrix <- aitch(data=example)
}
