% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBDMDB.R
\docType{data}
\name{allMGX}
\alias{allMGX}
\title{Microbiota of 107 individuals}
\format{A data frame with 1607 rows (patients) and 585 columns (species)}
\source{
\url{https://ibdmdb.org/tunnel/public/HMP2/WGS/1818/products}
}
\usage{
allMGX
}
\description{
Microbiota of 107 individuals
}
\keyword{datasets}
