% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{classify}
\alias{classify}
\title{SVM classifier}
\usage{
classify(data, y, kernel, p = 0.8, C = 1, G = 0, k, prob = FALSE,
  classimb = "no", type = "ubOver")
}
\arguments{
\item{data}{Input data}

\item{y}{Reponse variable (categoric)}

\item{kernel}{"cRBF" for clrRBF, "qJac" for quantitative Jaccard and  "wqJacc" for quantitative Jaccard with weights}

\item{p}{Proportion of total data instances in the training set}

\item{C}{A vector with the possible costs to evaluate via k-Cross-Val. If no argument is provided cross-validation is not performed.}

\item{G}{Gamma hyperparameter}

\item{k}{The k for the k-Cross Validation. Minimum k = 2.}

\item{prob}{if TRUE builds a model for calculating class probabilities}

\item{classimb}{"weights" to introduce class weights in the SVM algorithm and "data" to oversampling. If other arguments are provided nothing is done.}

\item{type}{Procedure to data oversampling ("ubOver","ubUnder" or "ubSMOTE")}
}
\value{
Confusion matrix
}
\description{
SVM classifier
}
\examples{
diag <- as.numeric(speMGX[,1])
diag[diag == 3] <- 1  # De 3 a 2 classes: No Malalt /  malalt
classify(data=speMGX[,7:ncol(speMGX)],diag,kernel="qJac",C=c(0.1,1),k=10)
classify(data=speMGX[,7:ncol(speMGX)],diag,kernel="qJac",C=1,classimb="data", type="ubOver")
}
