% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernFunct.R
\name{clrLin}
\alias{clrLin}
\title{Compositional Linear kernel}
\usage{
clrLin(data, cos.norm = TRUE)
}
\arguments{
\item{data}{A matrix or data.frame containing compositiona data in raw counts}

\item{cos.norm}{If TRUE the cosine normalization is applied}
}
\value{
The compositional linear kernel matrix
}
\description{
Compositional Linear kernel
}
\examples{
kmatrix <- clrLin(data=soil$abund)
}
