% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernFunct.R
\name{clrRBF}
\alias{clrRBF}
\title{Aitchison RBF kernel}
\usage{
clrRBF(data, h = NULL)
}
\arguments{
\item{data}{A matrix or data.frame containing compositiona data in raw counts}

\item{h}{Gamma hyerparameter. If NULL, the Aitchison distances are returned}
}
\value{
The Aitchison RBF kernel.
}
\description{
Aitchison RBF kernel
}
\examples{
kmatrix <- clrRBF(data=soil$abund,h=0.001)
}
