% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{CSSnorm}
\alias{CSSnorm}
\title{Cumulative Sum Scaling Normalization}
\usage{
CSSnorm(data, pcount = 0)
}
\arguments{
\item{data}{Input data}

\item{pcount}{Pseudocount value (default value 0)}
}
\value{
CSS normalized data
}
\description{
Cumulative Sum Scaling Normalization
}
\examples{
soilData <- CSSnorm(data=soil$abund)
}
