% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformMeasures.R
\name{F1}
\alias{F1}
\title{F1}
\usage{
F1(ct, min.class = 2)
}
\arguments{
\item{ct}{Confusion Matrix}

\item{min.class}{Minority class}
}
\value{
F1
}
\description{
F1
}
\examples{
model <- classify(data=soil$abund,y=soil$metadata[,"env_feature"],kernel="clin")
F1(model$"conf.matrix")
}
