% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBDMDB.R
\docType{data}
\name{FcalWeek}
\alias{FcalWeek}
\title{Fecal Calprotectin levels}
\format{A data frame with 131 rows (patients) and 52 columns (week number)}
\source{
\url{https://ibdmdb.org/tunnel/products/HMP2/Metadata/hmp2_metadata.csv}
}
\usage{
FcalWeek
}
\description{
Fecal Calprotectin levels
}
\keyword{datasets}
