% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kMatrixIntegr.R
\name{fuseData}
\alias{fuseData}
\title{Fuse Data}
\usage{
fuseData(DATA, kernel, y, coeff)
}
\arguments{
\item{DATA}{A list of the *m* data to fuse}

\item{kernel}{A vector of length *m* with the kernels to use in each data}

\item{y}{Only if "wqJac" is chosen: response variable}

\item{coeff}{A vector of length *m* with the weight of each kernel data,
or length(m) - 1 if the last coefficient is 1 - sum(coeff).
If absent,  all data is considered equally important.}
}
\value{
A consensus kernel matrix (via calling KInt)
}
\description{
Fuse Data
}
\examples{
d <- list()
d[[1]] <- matrix(abs(rnorm(20)),nrow=4,ncol=5)
d[[2]] <- matrix(abs(rnorm(20)),nrow=4,ncol=5)
fuseData(DATA=d,kernel=c("qJac","cRBF"))
fuseData(DATA=d,kernel=c("qJac","cRBF"),coeff=c(0.9,0.1))
}
