% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBDMDB.R
\docType{data}
\name{genMGX}
\alias{genMGX}
\title{Intestinal microbiota of 84 individuals}
\format{A data frame with 84 rows and 199 variables:}
\source{
\url{https://ibdmdb.org/tunnel/public/HMP2/WGS/1818/products}
}
\usage{
genMGX
}
\description{
Genus abundances of intestinal microbiota of 84 individuals, 41 with
Chron's Disease, 24 with Ulcerative Colitis, and 19 without inflamatory
inflammatory bowel disease.
}
\keyword{datasets}
