% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\docType{data}
\name{growth}
\alias{growth}
\title{Growth data}
\format{
A matrix with 2883 rows (93 individuals * 31 measures) and 2 variables:
 \describe{
  \item{Age}{Age in years}
  \item{Height}{Height in cm}
}
}
\source{
Berkeley Growth Study Data
Ramsay, James O., and Silverman, Bernard W. (2006), Functional Data Analysis, 2nd ed., Springer, New York.

Ramsay, James O., and Silverman, Bernard W. (2002), Applied Functional Data Analysis, Springer, New York, ch. 6.

Tuddenham, R. D., and Snyder, M. M. (1954) "Physical growth of California boys and girls from birth to age 18", University of California Publications in Child Development, 1, 183-364.
}
\usage{
growth
}
\description{
Longitudinal height data of 54 girls and 39 boys (93 individuals in total) from age 1 to 18.
}
\keyword{datasets}
