% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernFunct.R
\name{JSK}
\alias{JSK}
\title{This function returns the Jensen Shannon Kernel}
\usage{
JSK(data)
}
\arguments{
\item{data}{A matrix or data.frame containing nonnegative values.}
}
\value{
The JS kernel matrix
}
\description{
This function returns the Jensen Shannon Kernel
}
\examples{
example <- matrix(abs(rnorm(12)),nrow=4,ncol=3)
kmatrix <- JSK(data=example)
}
