% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised.R
\name{kernPCA}
\alias{kernPCA}
\title{kernel PCA}
\usage{
kernPCA(data, kernel, H = NULL, y, dim = c(1, 2), colors,
  na.col = "grey70", title, legend = TRUE, labels = FALSE)
}
\arguments{
\item{data}{Input data: a matrix or data.frame.}

\item{kernel}{"linear" for linear kernel, "rbf" for RBF, "cRBF" for clrRBF, "qJac" for quantitative Jaccard.
"matrix" if a pre-calculated kernel matrix is given as input.}

\item{H}{Kernel hyperparameter.}

\item{y}{Response vector; optional. Colors each dot (individual) according to its value in the response / target variable.}

\item{dim}{The two PC that have to be displayed. Defaults to the two first PC.}

\item{colors}{Dot fill color; optional.}

\item{na.col}{Depends on y and col. Dot fill color for the y missing values; defaults to grey.}

\item{title}{Plot title}

\item{legend}{Defaults to TRUE. A legend with the color corresponding to each y value.}

\item{labels}{If true, each dot will be labeled with its row number.}
}
\value{
A  k-PCA plot generated with ggplot2.
}
\description{
kernel PCA
}
\examples{
kernPCA(soilDataRaw[-89,],kernel="cRBF", H=0.0001,y=soilMetaData$ph[-89], col=c("aquamarine2","orchid3"),
title = "Soil kernel PCA",legend = TRUE)
}
