% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernFunct.R
\name{Kfun}
\alias{Kfun}
\title{Functional linear kernel}
\usage{
Kfun(data, domain, cos.norm = FALSE)
}
\arguments{
\item{data}{Output of lsq()}

\item{domain}{Domain}

\item{cos.norm}{If TRUE, perform cosine normalization}
}
\value{
A kernel matrix
}
\description{
Functional linear kernel
}
\examples{
growth2 <- growth[,2:3]
colnames(growth2) <-  c( "time", "height")
fitted <- lsq(data=growth2,degree=2)
Kfun(fitted,domain=c(11,18))
}
