% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kMatrixIntegr.R
\name{KInt}
\alias{KInt}
\title{Kernel Matrix Integration}
\usage{
KInt(data, coeff = "mean")
}
\arguments{
\item{data}{A three-dimensional ({n·d·m}) array containing {m} kernel matrices}

\item{coeff}{A vector of length {m} with the weight of each kernel matrix}
}
\value{
A kernel matrix
}
\description{
Kernel Matrix Integration
}
\examples{
DATA <- array(dim=c(4,4,3))
DATA[,,1] <- matrix(abs(rnorm(16)),nrow=4,ncol=4)
diag(DATA[,,1]) <- 1
DATA[,,2] <- matrix(abs(rnorm(16)),nrow=4,ncol=4)
diag(DATA[,,2]) <- 1
DATA[,,3] <- matrix(abs(rnorm(16)),nrow=4,ncol=4)
diag(DATA[,,3]) <- 1
KInt(data=DATA)
KInt(data=DATA,coeff=c(0.5,0.3,0.2))
}
