% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernFunct.R
\name{Linear}
\alias{Linear}
\title{Linear kernel}
\usage{
Linear(data, cos.norm = TRUE)
}
\arguments{
\item{data}{A matrix or data.frame with real numbers}

\item{cos.norm}{If TRUE the cosine normalization is applied}
}
\value{
The linear kernel matrix
}
\description{
Linear kernel
}
\examples{
example <- matrix(rnorm(12),nrow=4,ncol=3)
kmatrix <- Linear(data=example)
}
