% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBDMDB.R
\docType{data}
\name{longMGX}
\alias{longMGX}
\title{Longitudinal data (intestinal microbiota)}
\format{A list containing: a matrix with the species abundances of 107 individuals with 4 visits each one,
a vector with the number of individuals and visits, and a matrix with the distance among the visits (in weeks)}
\source{
\url{https://ibdmdb.org/tunnel/public/HMP2/WGS/1818/products}
}
\usage{
longMGX
}
\description{
Longitudinal data (intestinal microbiota)
}
\keyword{datasets}
