% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeries.R
\name{lsq}
\alias{lsq}
\title{Least Squares Coefficients}
\usage{
lsq(data, degree = 1, orthog = FALSE, scale = FALSE)
}
\arguments{
\item{data}{A matrix with longitudinal data, with a column named "time" containing the time
(in arbitrary units) of the measure. lsq() will consider that the row names that share id are repeated
measures at different times from the same individual.}

\item{degree}{The degree of the polynomial to be fitted.}

\item{orthog}{If true, use orthogonal polynomials.}

\item{scale}{If true, divide each coefficient by its variance}
}
\value{
An object of class lsq containing the intercept and coefficients over time of for each variable and individual.
}
\description{
Least Squares Coefficients
}
\examples{
growth2 <- growth[,2:3]
colnames(growth2) <-  c( "time", "height")
lsq(data=growth2,degree=2)
}
