% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kMatrixIntegr.R
\name{meanInt}
\alias{meanInt}
\title{Kernel Matrix Integration (via the mean)}
\usage{
meanInt(data)
}
\arguments{
\item{data}{A three-dimensional ({n·d·m}) array containing {m} kernel matrices}
}
\value{
The consensus kernel matrix
}
\description{
The most simple approach: mean of matrices of equal dimensions
}
\examples{
DATA <- array(dim=c(4,5,3))
DATA[,,1] <- matrix(abs(rnorm(20)),nrow=4,ncol=5)
DATA[,,2] <- matrix(abs(rnorm(20)),nrow=4,ncol=5)
DATA[,,3] <- matrix(abs(rnorm(20)),nrow=4,ncol=5)
intMatrix <- meanInt(data=DATA)
}
