% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBDMDB.R
\docType{data}
\name{metaLongMGX}
\alias{metaLongMGX}
\title{Longitudinal metadata (intestinal microbiota)}
\format{A table with the sampling time, individual diagnosis, sex, age etc}
\source{
\url{https://ibdmdb.org/tunnel/public/HMP2/WGS/1818/products}
}
\usage{
metaLongMGX
}
\description{
Longitudinal metadata (intestinal microbiota)
}
\keyword{datasets}
