% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeKernel.R
\name{minSep}
\alias{minSep}
\title{Minumum Separation Samples}
\usage{
minSep(table, n)
}
\arguments{
\item{table}{A n·d data.frame or matrix with n individuals and d ordered visits, stating if the individual came to the
visit (TRUE or 1) or no (FALSE or 0). The time must be specified as the column names.}

\item{n}{Number of samples}
}
\value{
A vector containing, for each individual, the name of the first sample of the minimum period
}
\description{
Minumum Separation Samples
}
\examples{
individuals <- which(rowSums(visitTable) > 3)
minSep(table=visitTable[individuals,],n=4) # The four samples with minimum separation
}
