% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformMeasures.R
\name{Prec}
\alias{Prec}
\title{Precision}
\usage{
Prec(ct, min.class = 2)
}
\arguments{
\item{ct}{Confusion Matrix}

\item{min.class}{Minority class}
}
\value{
Precision
}
\description{
Precision
}
\examples{
model <- classify(data=soil$abund,y=soil$metadata[,"env_feature"],kernel="clin")
Prec(model$"conf.matrix")
}
