% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernFunct.R
\name{RBFun}
\alias{RBFun}
\title{RBF kernel for functions}
\usage{
RBFun(data, h = NULL, domain)
}
\arguments{
\item{data}{Output of lsq()}

\item{h}{Gamma hyerparameter. If NULL, the L2 norms are returned}

\item{domain}{Domain}
}
\value{
The L2 norm (euclidean distance) between to matrices
}
\description{
RBF kernel for functions
}
\examples{
growth2 <- growth
colnames(growth2) <-  c( "time", "height")
fitted <- lsq(data=growth2,degree=2)
RBFun(fitted,domain=c(1,18),h=0.00001)
}
