% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformMeasures.R
\name{Rec}
\alias{Rec}
\title{Recall}
\usage{
Rec(ct, min.class = 2)
}
\arguments{
\item{ct}{Confusion Matrix}

\item{min.class}{Minority class}
}
\value{
Recall
}
\description{
Recall
}
\examples{
model <- classify(data=soil$abund,y=soil$metadata[,"env_feature"],kernel="clin")
Rec(model$"conf.matrix")
}
