% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoker.R
\docType{data}
\name{smoker}
\alias{smoker}
\title{Airway samples from smokers and nonsmokers}
\format{
A list containing:
 \describe{
  \item{abund}{Bacterial abundances}
  \item{metadata}{Sample information}
}
}
\source{
\url{https://qiita.ucsd.edu/study/description/103}
}
\usage{
smoker
}
\description{
Bacterial abundance in 29 smokers and 33 nonsmokers. Samples taken in quadruplicate:
left and right nasopharynx and oropharynx
}
\keyword{datasets}
