% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil.R
\docType{data}
\name{soil}
\alias{soil}
\title{89 samples of soil microbiota (raw counts)}
\format{
A list with
 \describe{
  \item{abund}{Bacterial abundances with 89 rows and 7396 taxa}
  \item{metadata}{Sample information}
  \item{taxonomy}{Taxonomy for each of the codes in $abund}
}
}
\source{
\url{https://qiita.ucsd.edu/study/description/103}
}
\usage{
soil
}
\description{
Bacterial abundances in 89 soils from across North and South America.
}
\keyword{datasets}
