% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil.R
\docType{data}
\name{soilData}
\alias{soilData}
\title{89 samples of soil microbiota (processed)}
\format{A data frame with 89 rows and 7396 variables:}
\source{
\url{https://qiita.ucsd.edu/study/description/103}
}
\usage{
soilData
}
\description{
89 samples of soil microbiota (processed)
}
\keyword{datasets}
