% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil.R
\docType{data}
\name{soilDataRaw}
\alias{soilDataRaw}
\title{89 samples of soil microbiota (raw counts)}
\format{A data frame with 89 rows and 7396 variables:}
\source{
\url{https://qiita.ucsd.edu/study/description/103}
}
\usage{
soilDataRaw
}
\description{
Bacterial abundances in 89 soils from across North and South America.
}
\keyword{datasets}
