% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil.R
\docType{data}
\name{soilMetaData}
\alias{soilMetaData}
\title{89 samples of soil microbiota (metadata)}
\format{A data frame with 89 rows and 30 variables:}
\source{
\url{https://qiita.ucsd.edu/study/description/103}
}
\usage{
soilMetaData
}
\description{
89 samples of soil microbiota (metadata)
}
\keyword{datasets}
