% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeKernel.R
\name{TimeK}
\alias{TimeK}
\title{Time Kernel Matrix}
\usage{
TimeK(data, h)
}
\arguments{
\item{data}{A n·d data.frame or matrix with n individuals and d ordered visits, stating if the individual came to the
visit (TRUE or 1) or no (FALSE or 0). The time must be specified as the column names.}

\item{h}{If argument, the functions returns exp(-h * |tj-ti|)}
}
\value{
A time matrix
}
\description{
Time Kernel Matrix
}
\examples{
DATA <- matrix(c(rep(1,5),rep(0,4),rep(1,4),rep(0,4),1,0,0,1,1,0,1,1,rep(0,4),1),ncol=6,nrow=5)
colnames(DATA) <- c(0:3,5,9) # Measured in weeks
rownames(DATA) <- paste("Indiv",1:5,sep="")
TimeK(data=DATA)
}
