% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBDMDB.R
\docType{data}
\name{visitTable}
\alias{visitTable}
\title{Metagenome sampling times of 131 individuals (in weeks)}
\format{A data frame with 131 rows (patients) and 55 columns (week number)}
\source{
\url{https://ibdmdb.org/tunnel/public/HMP2/WGS/1818/products}
}
\usage{
visitTable
}
\description{
Metagenome sampling times of 131 individuals (in weeks)
}
\keyword{datasets}
