% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernFunct.R
\name{wqJacc}
\alias{wqJacc}
\title{Quantitative Jaccard with weights}
\usage{
wqJacc(data, w, y, h = NULL)
}
\arguments{
\item{data}{A matrix or data.frame containing nonnegative values.}

\item{w}{a vector of weights as long as the number of variables.
Class should be set to "numeric". If empty, weights will be authomathically computed as the RF mean
decrease in Gini index (see below)}

\item{y}{If a vector of weights is not provided, a vector of responses for computing the RF
weights should be provided instead.}

\item{h}{An hyperparametr}
}
\value{
The weighted quantitative Jaccard kernel matrix
}
\description{
This function delivers the quantitative Jaccard kernel matrix, also known as Ruzicka similarity,
with the option to weight each variable.
}
\examples{
example <- matrix(abs(rnorm(12)),nrow=4,ncol=3)
weights <- c(0.1,0.7,0.2)
qJaccMatrix <- wqJacc(data=example,w = weights)
}
