% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformMeasures.R
\name{Acc}
\alias{Acc}
\title{Accuracy}
\usage{
Acc(ct)
}
\arguments{
\item{ct}{Confusion Matrix}
}
\value{
Accuracy
}
\description{
Accuracy
}
\examples{
model <- classify(data=soil$abund,y=soil$metadata[,"env_feature"],kernel="clin")
Acc(model$"conf.matrix")
}
