% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier.R
\name{classify}
\alias{classify}
\title{SVM classifier}
\usage{
classify(data, y, classes = 2, kernel, p = 0.8, C, k = 10,
  classimb = "no", type = "ubOver")
}
\arguments{
\item{data}{Input data}

\item{y}{Reponse variable (categoric)}

\item{classes}{Number of classes (2 or 3)}

\item{kernel}{"qJac" for quantitative Jaccard and "wqJacc" for quant Jaccard with weights}

\item{p}{Proportion of total data instances in the training set}

\item{C}{A vector with the possible costs to evaluate via k-Cross-Val. If no argument is provided cross-validation is not performed.}

\item{k}{The k for the k-Cross Validation. Minimum k = 2.}

\item{classimb}{"weights" to introduce class weights in the SVM algorithm and "data" to oversampling. If other arguments are provided nothing is done.}

\item{type}{Procedure to data oversampling ("ubOver" or "ubSMOTE")}
}
\value{
Confusion matrix
}
\description{
SVM classifier
}
\examples{
classify(data=speMGX[,7:ncol(speMGX)],speMGX[,1],kernel="qJac",C=c(0.1,1),k=10)
classify(data=speMGX[,7:ncol(speMGX)],speMGX[,1],kernel="qJac",C=1,classimb="data", type="ubOver")
}
