% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{clr}
\alias{clr}
\title{clr transform}
\usage{
clr(data)
}
\arguments{
\item{data}{A matrix or data.frame with compositional data (raw counts)}
}
\value{
clr transformation of data
}
\description{
clr transform
}
\examples{
clr(data=soil$abund)
}
