% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernFunct.R
\name{clrRBF}
\alias{clrRBF}
\title{clr transf + RBF kernel}
\usage{
clrRBF(data, h)
}
\arguments{
\item{data}{A matrix or data.frame containing only positive values.}

\item{h}{Kernel hyperparameter (gamma)}
}
\value{
RBF kernel over a clr transformated data
}
\description{
clr transf + RBF kernel
}
\examples{
example <- matrix(abs(rnorm(12)),nrow=4,ncol=3)
kernelMatrix <- clrRBF(data=example)
}
