% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{CSSnorm}
\alias{CSSnorm}
\title{Cumulative Sum Scaling Normalization}
\usage{
CSSnorm(data)
}
\arguments{
\item{data}{Input data}
}
\value{
Normalized data
}
\description{
Cumulative Sum Scaling Normalization
}
\examples{
soilData <- CSSnorm(data=soil$abund)
}
