% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalization.R
\name{CSSnorm}
\alias{CSSnorm}
\title{DATA NORMALIZATION}
\usage{
CSSnorm(data)
}
\arguments{
\item{data}{Input data}
}
\value{
Normalized data
}
\description{
DATA NORMALIZATION
}
\examples{
soilData <- CSSnorm(data=t(soilDataRaw))
}
