% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformMeasures.R
\name{F1}
\alias{F1}
\title{F1}
\usage{
F1(ct)
}
\arguments{
\item{ct}{Confusion Matrix}
}
\value{
F1
}
\description{
F1
}
\examples{
diag <- as.numeric(speMGX[,1])
diag[diag == 3] <- 1  # De 3 a 2 classes: No Malalt /  malalt
CM <- classify(data=speMGX[,7:ncol(speMGX)],y=diag,kernel="qJac")
F1(CM)
}
