% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unsupervised.R
\name{hklust}
\alias{hklust}
\title{Hierarchical clustering}
\usage{
hklust(data, kernel = "linear", H = NULL, method = "ward.D2",
  plot = TRUE, labels = NULL, title = NULL, cut = NULL,
  height = NULL, colors = "black")
}
\arguments{
\item{data}{Input data: a matrix or data.frame.}

\item{kernel}{"linear" for linear kernel, "rbf" for RBF, "cRBF" for clrRBF, "qJac" for quantitative Jaccard.
"matrix" if a pre-calculated kernel matrix is given as input.}

\item{H}{Kernel hyperparameter.}

\item{method}{the agglomeration method to be used: "ward.D", "ward.D2" (Ward's (1963) clustering criterion),
"single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC).}

\item{plot}{TRUE to return the plot}

\item{labels}{A vector of labels for the leaves of the tree. Row names of
the input data are used by default. If set to FALSE, no labels at all are plotted.}

\item{title}{Plot title}

\item{cut}{an integer scalar or vector with the desired number of groups}

\item{height}{numeric scalar or vector with heights where the tree should be cut}

\item{colors}{border color(s) for the rectangles.}
}
\value{
An object of class hclust with or without a plot of the cluster dendogram
}
\description{
Hierarchical clustering
}
\examples{
hklust(soilDataRaw[-89,],kernel="cRBF", title = "Soil data cluster dendogram",cut=3,colors=2:4)
}
