% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeKernel.R
\name{lsq}
\alias{lsq}
\title{Least Squares Coefficients}
\usage{
lsq(data, degree)
}
\arguments{
\item{data}{A matrix with longitudinal data, with a column named "time" containing the time
(in arbitrary units) of the measure. lqs() will consider that the row names that share id are repeated
measures at different times from the same individual.}

\item{degree}{The degree of the polynomial to be fitted.}
}
\value{
A matrix with the standardized intercept and coefficients of the model for each individual.
}
\description{
Least Squares Coefficients
}
\examples{
colnames(growth) <-  c("sex", "time", "height")
lsq(data=growth[,2:3],degree=2)
}
