% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutDetect.R
\name{outliers}
\alias{outliers}
\title{SVM outlier detection}
\usage{
outliers(data, kernel, nu)
}
\arguments{
\item{data}{Input data}

\item{kernel}{"qJac" for quantitative Jaccard and "wqJacc" for quant Jaccard with weights}

\item{nu}{Hyperparameter nu}
}
\value{
The indexes of the outliers
}
\description{
SVM outlier detection
}
\examples{
outliers(data=MGXdata[,7:ncol(MGXdata)],kernel="qJac",nu=0.9)
}
