% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutDetect.R
\name{outliers}
\alias{outliers}
\title{SVM outlier detection / one class SVM}
\usage{
outliers(data, y, kernel, nu, p = 0.8, k, H = 0)
}
\arguments{
\item{data}{Input data}

\item{y}{Reponse variable. If a value is provided, outliers() functions as an one-class SVM.}

\item{kernel}{"qJac" for quantitative Jaccard and "wqJacc" for quant Jaccard with weights.
"matrix" if a pre-calculated kernel matrix is given as input.}

\item{nu}{Hyperparameter nu}

\item{p}{If a value for y is provided, p is the proportion of total data instances in the training set}

\item{k}{The k for the k-Cross Validation. Minimum k = 2.}

\item{H}{Hyperparameter gamma}
}
\value{
The indexes of the outliers (outlier detection) or, if a value is provided for y, the confusion matrix (one-class SVM)
}
\description{
outliers() has two principal usages: unsupervised detection of outliers in data, o supervised one-class SVC.
}
\details{
If outliers() is used in a supervised way and the input data has repeated rownames, classify() will consider that the row names that share id are repeated
measures from the same individual. The function will ensure that all repeated measures are used either to train
or to test the model, but not for both, thus preserving the independence between the training and tets sets.
}
\examples{
# Outlier detection
outliers(data=soilDataRaw,kernel="cRBF",nu=0.3)
## One-class SVM:
## Preparing the y
diag <- as.numeric(speMGX[,1])
diag[diag == 3] <- 1  # De 3 a 2 classes: No Malalt /  malalt
diag[diag == 2] <- 0
diag <- as.numeric(!diag) # No malalt classe 1, malalt classe 0
## One-class SVM changing the percentage of data for training (70\%) and the hyperparameter nu:
outliers(data=speMGX[,7:ncol(speMGX)],y=diag,kernel="qJac",nu=0.2,p=0.7)
## One-class SVM with 10-Cross-Validation:
outliers(data=speMGX[,7:ncol(speMGX)],y=diag,kernel="qJac",nu=c(0.45,0.5),H=c(0.1,1),k=10)
}
