% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PerformMeasures.R
\name{Prec}
\alias{Prec}
\title{Precision}
\usage{
Prec(ct, min.class = 2)
}
\arguments{
\item{ct}{Confusion Matrix}

\item{min.class}{Minority class}
}
\value{
Precision
}
\description{
Precision
}
\examples{
diag <- as.numeric(speMGX[,1])
diag[diag == 3] <- 1  # De 3 a 2 classes: No Malalt /  malalt
CM <- classify(data=speMGX[,7:ncol(speMGX)],y=diag,kernel="qJac")
Prec(CM$"conf.matrix")
}
