% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernFunct.R
\name{qJacc}
\alias{qJacc}
\title{Quantitative Jaccard}
\usage{
qJacc(data, h)
}
\arguments{
\item{data}{A matrix or data.frame containing nonnegative values.}

\item{h}{An hyperparametr}
}
\value{
The quantitative Jaccard kernel matrix
}
\description{
This function delivers the quantitative Jaccard kernel matrix, also known as Ruzicka similarity.
}
\examples{
example <- matrix(abs(rnorm(12)),nrow=4,ncol=3)
qJaccMatrix <- qJacc(data=example)
}
