% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{regress}
\alias{regress}
\title{SVM regression}
\usage{
regress(data, y, kernel, p = 0.8, C = 1, G = 0, E = 0.1, k)
}
\arguments{
\item{data}{Input data}

\item{y}{Reponse variable (continuous)}

\item{kernel}{"cRBF" for clrRBF, "qJac" for quantitative Jaccard and  "wqJacc" for quantitative Jaccard with weights}

\item{p}{Proportion of total data instances in the training set}

\item{C}{A vector with the possible costs to evaluate via k-Cross-Val. If no argument is provided cross-validation is not performed.}

\item{G}{Gamma hyperparameter}

\item{E}{Epsilon hyperparameter}

\item{k}{The k for the k-Cross Validation. Minimum k = 2.}
}
\value{
NMSE (normalized mean squared error)
}
\description{
SVM regression
}
\examples{
regress(data=soilDataRaw[-89,],soilMetaData$ph[-89],kernel="cRBF",C=c(1,10,50),k=10)
}
