% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.R
\name{regress}
\alias{regress}
\title{SVM regression}
\usage{
regress(data, y, coeff, kernel, p = 0.8, C = 1, H = NULL, E = 0.1,
  k)
}
\arguments{
\item{data}{Input data}

\item{y}{Reponse variable (continuous)}

\item{coeff}{ONLY IN MKL CASE: A *t·m* matrix of the coefficients, where *m* are the number of different data types and *t* the number of
different coefficient combinations to evaluate via k-CV.}

\item{kernel}{"cRBF" for clrRBF, "qJac" for quantitative Jaccard and  "wqJacc" for quantitative Jaccard with weights.
"matrix" if a pre-calculated kernel matrix is given as input.}

\item{p}{Proportion of total data instances in the training set}

\item{C}{A vector with the possible costs to evaluate via k-Cross-Val. If no argument is provided cross-validation is not performed.}

\item{H}{Gamma hyperparameter}

\item{E}{Epsilon hyperparameter}

\item{k}{The k for the k-Cross Validation. Minimum k = 2.}
}
\value{
NMSE (normalized mean squared error)
}
\description{
regress() automatically trains a Support Vector Regrssion model, tests it and returns the Normalized Mean Squared Error.
}
\details{
If the input data has repeated rownames, classify() will consider that the row names that share id are repeated
measures from the same individual. The function will ensure that all repeated measures are used either to train
or to test the model, but not for both, thus preserving the independence between the training and tets sets.
}
\examples{
# Data normalization
soilData <- CSSnorm(data=t(soilDataRaw[-89,]))
# Simple regression without tuning the hyperparameters
regress(data=soilData,soilMetaData$ph[-89],kernel="qJac")
# The percentage of data for training can be changed (default: 0.8 Training / 0.2 Test):
regress(data=soilData,soilMetaData$ph[-89],kernel="qJac",p=0.6)
# Regression with 10-Cross-Validation to choose the best Cost and Epsilon:
regress(data=soilData, y=soilMetaData$ph[-89], kernel="qJac", C=c(0.1,1,10), E = c(0.01,0.1), k=10)
}
